+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `skip_if_no_field` and `invert` options to the regex used in the generic
   filter.
 - Basic KiCost support.
+- Experimental mechanism to change 3D models according to the variant.
 
 ### Changed
 - Errors and warnings from KiAuto now are printed as errors and warnings.